Consolidate all API calls within TMF8801_Control
